Added ref frame transformations, droop controller, and DQ power calcs
The reference frames now include three-phase ABC, stationary alpha-beta,
synchronous DQZ, and Polar. New traits are introduced and implemented for each reference frame to transform between the frames.

Created and tested the droop GFM controller object. Note the controller
is not per-unitized at this moment.

In addition to the existing alpha-beta reference frame power calculation
function, added a similar calculation function for the DQZ reference
frame as well as seperate functions for calculating just the active or
reactive powers seperately.