Update gfm controller w_nom param to be f_nom
This change allows for a smaller integer bit for the fixed-point number
used for the gfm controller. Allows for up to 24-bit fractional (I8F24),
which provides good performance for the dVOC controller, and fair
performance with the droop controller. Seeing oscillations on the
voltage and powers still, but this may be due to lossy fixed-point
conversions in the simulation.