Added GFM object with presyc functionality
Combined grid-forming controllers into one file, 'gfm.rs'.

Created a GFM struct that contains a grid-forming controller object and
presync parameters and implemented a gfm_step function to step the GFM
object with either normal dynamics or presync dynamics.

Made a trait for a GFMInterface that allows the user to get the voltage
and set the power references of the GFM object's controller.