Implemented line trait for lines, and Fixed LineToBus Dynamics
The line trait has been implemented for the RlBranch and LclFilter
objects. Additionally, a switch state was added to both of these structs
and functions to interact with this switch state were added to the line
trait.

The LineToBus object had issues with its dynamics which were due to the
indexing of the nalgebra matrixes in the dynamics function. This
indexing was fixed and tested with dVOC control and an RL branch in a
SMSB simulation.