Renamed files, Added PLLs, Added OrthSigGen, Renamed GFM
Renamed files with abbreviations to the full word for clarity.

Renamed the GFM object to PresyncInvController to clarify what it does
as it could technically be used for any polar coordinate inverter
controller.

Added files (dynamics, inverter) and moved blocks of code that could
stand alone (traits and presync object).

Added SRF and DSRF PLL objects and implemented their dynamics.

Added a SOGI orthogonal signal generator for single-phase controllers.

Note that the PLLs and orthogonal signal generator still need to be validated.