Added wrap_angle to ACVS and moved OSG definition to new osg file
Added a function to the ACVS object to easily wrap its angle.

Moved orthogonal signal generator definitions to a new file / renamed
the single_phase.rs file.

Defined a single-phase line, but this will likely be unused as the
alpha-beta definition contains more information for a simulation and can
represent a single-phase line itself.